Display banner if dataset has customDownloadTab attribute
